requirements.txt
matplotlib
matplotlib-inline
numpy
pandas
pillow
requests
seaborn
tokenizers
transformers
pyngrok
streamlit
streamlit_ace
joblib
tensorflow
keras
wikipedia
openai
gradio
tf-keras
torch
torchvision
google-api-python-client
python-dotenv
google
lxml
sklearn
underthesea
synonyms
sentence_transformers
